enable strict typechecking
i turned on the option for strict typechecking. fields that can be
undefined (for example, @Input fields on components) have been
explicitly marked as optional, and non-null assertions have been added
in places where they are used.

also, I've changed all uses of null to be undefined instead, so that
we only have to deal with one form of missing value and not two.